backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==1.0.0
google-cloud-aiplatform==1.70.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
websockets==13.1
python-multipart==0.0.9
PyMuPDF==1.24.13
httpx==0.27.2
python-dotenv==1.0.1
